requirements.txt
asgiref==3.10.0
certifi==2025.8.3
distlib==0.4.0
Django==5.2.8
django-debug-toolbar==6.1.0
filelock==3.19.1
mysqlclient==2.2.7
packaging==25.0
pillow==12.0.0
pipenv==2025.0.4
platformdirs==4.4.0
setuptools==80.9.0
sqlparse==0.5.3
tzdata==2025.2
virtualenv==20.34.0
